api/requirements.txt
# FastAPI core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# gRPC client
grpcio==1.60.0
grpcio-tools==1.60.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# JWT
PyJWT[crypto]==2.8.0
cryptography==42.0.0

# Security
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# HTTP client for external APIs
httpx==0.26.0

# AWS SDK (for S3 presigned URLs)
boto3==1.34.22

# Stripe SDK
stripe==8.0.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For TestClient

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
